Move fake AI client into test_infra
