+++ pkgs/flutter_genui/test/test_infra/fake_ai_client.dart
